reduce warnings for pedantic compiler options
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk@1009 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ qwt/src/qwt_compass_rose.cpp
@@ -155,7 +155,7 @@ void QwtSimpleCompassRose::drawRose(
 
     for ( int j = 1; j <= numThornLevels; j++ )
     {
-        double step =  qPow( 2.0, j ) * M_PI / ( double )numThorns;
+        double step =  qPow( 2.0, j ) * M_PI / numThorns;
         if ( step > M_PI_2 )
             break;
